refactor(driver): extract output-path resolution into a tested seam
The opt driver derived the <stem>_optimized.<ext> sibling path and enforced
the never-optimize-in-place guard (including the hard-link identity check)
inline in main.rs, tested only through the giant cli_helper_tests module. Lift
those rules into a pure seam, src/output_path.rs, with a single public entry
point (resolve_output_path) and the derive/same-file helpers kept private.

Fix a reachable panic in the derive step while extracting: optimized_output_path
unwrapped file_stem()/to_str(), so 's11 opt' on an input with no file stem (.., /)
or a non-UTF-8 filename (arbitrary bytes on Linux) aborted the process instead of
reporting an error. The derive step is now fallible and resolve_output_path
propagates the error, which the driver already turns into a clean 'Error: ...' exit.

TDD: added red tests asserting resolve_output_path returns Err (not panic) for
stem-less and non-UTF-8 inputs, confirmed they panicked against the old code, then
made the derive fallible to turn them green. The four existing path tests move into
the module unchanged.

Register the module in lib.rs and add the Output-path resolution glossary entry to
CONTEXT.md, mirroring the disassembly-listing seam (#759).
